britecore/requirements.txt
alembic==1.0.5
arrow==0.12.1
bcrypt==3.1.4
binaryornot==0.4.4
blinker==1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cookiecutter==1.6.0
dominate==2.3.5
environs==4.0.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.4.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-Webpack==0.1.0
Flask-WTF==0.14.2
future==0.17.1
gunicorn==19.9.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jinja2-time==0.2.0
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.16.3
poyo==0.4.2
psycopg2==2.7.5
pycparser==2.19
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
requests==2.20.1
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.24.1
visitor==0.1.3
Werkzeug==0.14.1
whichcraft==0.5.2
WTForms==2.2.1
